Refactoring to change the way the underlying raw message data is organised.
Seperated out the header data from the body data, created a class to manage this data. For new mesages have the option of supplying extra body data (if required), however this is not required (see the status request/response messages.

Added better tests to exercise messages that are more than just the header to fully test all edge cases in the TcpSession operations. This refactor was a result of these testing changes which caused errors that could not be resolved given the current structure.